Home page skeleton is functional, removed sidebar
